Refactor snowplow directory to be utilities - thought this would be necessary if new slack and service now classes needed - at this point, this does not appear to be necessary
