+++ api/app/resources/service_requests_list.py
@@ -19,7 +19,7 @@ from qsystem import api, api_call_with_retry, db, oidc, socketio
 from app.models import Citizen, CitizenState, CSR, Period, PeriodState, Service, ServiceReq, SRState
 from app.schemas import CitizenSchema, ServiceReqSchema
 from marshmallow import ValidationError
-from ..snowplow.snowplow import SnowPlow
+from ..utilities.snowplow import SnowPlow
 
 @api.route("/service_requests/", methods=["POST"])
 class ServiceRequestsList(Resource):
